Refactor Opcode_Register to Register conversion
Registers appearing in opcodes use a 3-bit signature, which does not
match the order of the actual registers (A is last, and odd, and the
(HL) opcode refers to a memory address rather than a register). Added
fake registers so that a simple `getRegister_from_OpcodeRegister()`
function could be created, which will allow uncorrupted read/write
using an opcode register signature. This also simplifies the `LD` op
(and eventually others), requiring only a split between (HL) and
registers.

Debug functionality is added here to check for fake writes in debug
builds.