Add ADC function to simplify basic arithmetic
ADD can be implemented in terms of ADC with Carry cleared. SBC can be
implemented as ADC with flipped Carry and complemented addend. Finally
SUB can be implemented in terms of SBC with Carry cleared (although
because SBC flips Carry, we set Carry, and then ADC(~addend) as per SBC